requirements.txt
pytest>=8.0
python-telegram-bot[job-queue]>=21.0
python-dotenv>=1.0
